# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BE Nhựa API",
        "description": "API quản lý sản phẩm và biến thể sản phẩm",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/variant-groups": {
            "get": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Danh sách nhóm biến thể",
                "operationId": "5e805b0c79871e46ccc69b6ceb9c1100",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Tạo nhóm biến thể",
                "operationId": "d803a89e403d38064eda666c424f0655",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreVariantGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đã tạo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VariantGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/variant-groups/{id}": {
            "get": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Chi tiết nhóm biến thể",
                "operationId": "1f8a1b006400fe613dfb03cbb93b816b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VariantGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tồn tại"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Cập nhật nhóm biến thể",
                "operationId": "a3410c05b5a27c14bcf8c1bdff38dfe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVariantGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đã cập nhật",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VariantGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tồn tại"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Xóa nhóm biến thể",
                "operationId": "2e7fe85c1a2a057203b1c14532d09449",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã xóa"
                    },
                    "404": {
                        "description": "Không tồn tại"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreVariantGroupRequest": {
                "required": [
                    "group_code",
                    "group_name"
                ],
                "properties": {
                    "group_code": {
                        "type": "string",
                        "example": "color",
                        "maxLength": 100
                    },
                    "group_name": {
                        "type": "string",
                        "example": "Màu sắc",
                        "maxLength": 255
                    },
                    "option_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateVariantGroupRequest": {
                "properties": {
                    "group_code": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "group_name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "option_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
            "VariantGroupResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "group_code": {
                        "type": "string"
                    },
                    "group_name": {
                        "type": "string"
                    },
                    "sort_order": {
                        "description": "Chỉ có khi group thuộc product",
                        "type": "integer",
                        "nullable": true
                    },
                    "is_required": {
                        "description": "Chỉ có khi group thuộc product",
                        "type": "boolean",
                        "nullable": true
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VariantOptionResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "VariantOptionResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "variant_group_id": {
                        "type": "integer"
                    },
                    "option_code": {
                        "type": "string"
                    },
                    "option_name": {
                        "type": "string"
                    },
                    "sort_order": {
                        "type": "integer"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer token",
                "scheme": "bearer"
            }
        }
    },
...
}